feat(storeadmin): refactoring code
