+++ src/components/admin/users/AddStoreAdminDialog.tsx
@@ -1,0 +1,137 @@
+"use client";
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { Plus } from "lucide-react";
+import { CreateStoreAdminData } from "./typesAndInterfaces";
+
+interface AddStoreAdminDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  stores: { id: string; name: string; city: string; province: string }[];
+  formData: CreateStoreAdminData;
+  setFormData: (data: CreateStoreAdminData) => void;
+  onSubmit: (e: React.FormEvent) => void;
+  isSubmitting: boolean;
+}
+
+export function AddStoreAdminDialog({
+  open,
+  onOpenChange,
+  stores,
+  formData,
+  setFormData,
+  onSubmit,
+  isSubmitting,
+}: AddStoreAdminDialogProps) {
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogTrigger asChild>
+        <Button className="flex items-center space-x-2">
+          <Plus className="h-4 w-4" />
+          <span>Add Store Admin</span>
+        </Button>
+      </DialogTrigger>
+      <DialogContent className="sm:max-w-[425px]">
+        <form onSubmit={onSubmit}>
+          <DialogHeader>
+            <DialogTitle>Create New Store Admin</DialogTitle>
+            <DialogDescription>
+              Create a new store administrator account.
+            </DialogDescription>
+          </DialogHeader>
+
+          <div className="grid gap-4 py-4">
+            <div className="grid gap-2">
+              <Label htmlFor="name">Full Name</Label>
+              <Input
+                id="name"
+                value={formData.name}
+                onChange={(e) =>
+                  setFormData({ ...formData, name: e.target.value })
+                }
+                placeholder="Enter full name"
+                required
+              />
+            </div>
+            <div className="grid gap-2">
+              <Label htmlFor="email">Email</Label>
+              <Input
+                id="email"
+                type="email"
+                value={formData.email}
+                onChange={(e) =>
+                  setFormData({ ...formData, email: e.target.value })
+                }
+                placeholder="Enter email address"
+                required
+              />
+            </div>
+            <div className="grid gap-2">
+              <Label htmlFor="password">Password</Label>
+              <Input
+                id="password"
+                type="password"
+                value={formData.password}
+                onChange={(e) =>
+                  setFormData({ ...formData, password: e.target.value })
+                }
+                placeholder="Enter password"
+                required
+              />
+            </div>
+            <div className="grid gap-2">
+              <Label htmlFor="store">Assign Store</Label>
+              <Select
+                value={formData.storeId}
+                onValueChange={(value) =>
+                  setFormData({ ...formData, storeId: value })
+                }
+              >
+                <SelectTrigger>
+                  <SelectValue placeholder="Select a store" />
+                </SelectTrigger>
+                <SelectContent>
+                  {stores.map((store) => (
+                    <SelectItem key={store.id} value={store.id}>
+                      {store.name} - {store.city}, {store.province}
+                    </SelectItem>
+                  ))}
+                </SelectContent>
+              </Select>
+            </div>
+          </div>
+
+          <DialogFooter>
+            <Button
+              type="button"
+              variant="outline"
+              onClick={() => onOpenChange(false)}
+            >
+              Cancel
+            </Button>
+            <Button type="submit" disabled={isSubmitting}>
+              {isSubmitting ? "Creating..." : "Add Store Admin"}
+            </Button>
+          </DialogFooter>
+        </form>
+      </DialogContent>
+    </Dialog>
+  );
+}
